// lib/types.ts
export type CustomUser = NextAuthUser & PrismaUser;

export interface CreateUserEmailProps {
	user: {
		name: string | null | undefined;
		email: string | null | undefined;
	};
}